Work on the "select" block (1).
html - block "select", got the main classes, tags and part of the content. Two basic classes were created to work with the block. In the course of work, the block underwent minor changes.

css/scss - block "select" and its classes, were added to the style file.

img - images were added to work with the "select" block.