Work on the "select" block (3) + add JS animations.
html - the script "wow.min.js" was added to the html file, which is responsible for animation when scrolling the page. The necessary elements on the site, the properties of this library were used.

css/scss - for the block "select" and its classes the css style properties and flex-elements were used.

js - the script "wow.min.js" was added.